Fix: Don't crash if a message has no author
When setting the status message, the bot looks for the prior message, to
update it if it is found. To do this, it filters the message list by
author id. However, sometimes a message does not have an author (unsure
when; maybe when the author has left the server?), and the bot crashes.

This slipped by for a long time because the types don't mark the author
as nullable (although the api reference does).