Embed a snippet from pasted playground links
It's nice that playground links are shortened. It's not as nice when
someone shares only a playground link without any inline code samples
and you have to click to find out what their problem is.

Rather than ask people to share a playground link *and* a sample, we can
instead extract a snippet from the playground link and include it in the
embed. As an added bonus, embeds have a darker background color so
comments are higher-contrast and not quite as unreadable.

Playground links already include the selection line numbers in the url
parameters (albeit in an unstable sort order and 1-indexed), so if
someone shares a link with the selection included, use that for the
sample (capped at ~512 characters). Otherwise, use the first ~256 chars.
Ideally we'd run twoslash on the whole thing and then print out the
error and some context, but in practice twoslash is way too slow.

It's *around* those numbers because actually it finds the line which
would be truncated and includes the full thing instead. So yes, someone
could be a smarty-pants and cram everything on one line to get it all
embedded (still truncated at the actual discord message length limit).

For formatting, code blocks in embeds have only 56 characters before
they wrap. Most people use longer lines than that, so most anything
looks terrible when embedded. To fix this, run prettier on the sample
that we want to embed, trying to keep lines to 55 characters. It works
surprisingly well.