Fix: don't truncate playground selection mid-line
Commit 6f6567a196e9d9cf079fb15899a0bc5f79ef80fc describes the intended
behavior — if the max playground embed length is in the middle of a
line, include the full line.

This was the actual behavior for the default length, if nothing was
selected, but I'd forgotten to switch the selection behavior to this as
well; it was still being hard-truncated at 512 chars.

(There is still a hard-truncation at 2048 chars, Discord's max).